requirements-dev.txt
-r requirements.txt
django-debug-toolbar==4.4.6
ipython==8.29.0